user_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
redis==5.0.1
django-redis==5.4.0
requests==2.31.0